:recycle: Small refactors to config and comments
