+++ aksel.nav.no/website/app/api/markdown/routes/maler.ts
@@ -3,7 +3,7 @@ import {
   TEMPLATES_BY_SLUG_MARKDOWN_QUERY,
 } from "@/app/_sanity/queries";
 import type { ALL_TEMPLATES_MARKDOWN_QUERY_RESULT } from "@/app/_sanity/query-types";
-import { createRoute } from "./create-route";
+import { createRoute } from "../helpers/create-route";
 
 type MalerItem = ALL_TEMPLATES_MARKDOWN_QUERY_RESULT[number];
